requirements.txt
python-dotenv
sqlalchemy
pymysql
pyodbc
pandas
tqdm